+++ src/micapolos/tata8/Random.java
@@ -3,11 +3,11 @@ package micapolos.tata8;
 public class Random {
   private Random() {}
 
-  public static int between(int min, int max) {
-    return (int) Math.round(min + Math.random() * (max - min));
+  public static float between(float min, float max) {
+    return Math.round(min + Math.random() * (max - min));
   }
 
-  public static int until(int limit) {
+  public static float until(float limit) {
     return between(0, limit - 1);
   }
 }
